+++ andes/models/exciter.py
@@ -191,8 +191,8 @@ class EXDC2Model(ExcBase):
 
         self.vref0 = ConstService(info='Initial reference voltage input',
                                   tex_name='V_{ref0}',
-                                  v_str='vb0 + v')
-
+                                  v_str='vb0 + v',
+                                  )
         self.Se = Algeb(info='Saturation output',
                         tex_name='S_e',
                         unit='p.u.',
@@ -203,7 +203,9 @@ class EXDC2Model(ExcBase):
                         tex_name='V_p',
                         unit='p.u.',
                         v_str='vf0',
-                        e_str='(LA_x - KE * vp - Se * vp) / TE')
+                        e_str='(LA_x - KE * vp - Se * vp)',
+                        t_const=self.TE,
+                        )
 
         self.LS = Lag(u=self.v, T=self.TR, K=1.0, info='Sensing lag TF')