+++ andes/routines/eig.py
@@ -2,7 +2,7 @@ import logging
 import scipy.io
 
 from math import ceil, pi
-from cvxopt import mul, div
+from cvxopt import mul, div, spdiag
 from cvxopt.lapack import gesv
 
 from andes.io.txt import dump_data
@@ -32,19 +32,36 @@ class EIG(BaseRoutine):
         self.part_fact = None
 
     def calc_state_matrix(self):
-        """
-        Return state matrix and store to ``self.As``
+        r"""
+        Return state matrix and store to ``self.As``.
+
+        Notes
+        -----
+        For systems with the form
+
+        .. math ::
+
+            T \dot{x} = f(x, y) \\
+            0 = g(x, y)
+
+        The state matrix is calculated from
+
+        .. math ::
+
+            A_s = T^{-1} (f_x - f_y * g_y^{-1} * g_x)
 
         Returns
         -------
-        matrix
+        cvxopt.matrix
             state matrix
         """
         system = self.system
+
         gyx = matrix(system.dae.gx)
         self.solver.linsolve(system.dae.gy, gyx)
 
-        self.As = matrix(system.dae.fx - system.dae.fy * gyx)
+        inv_zf = spdiag((1 / system.dae.zf).tolist())
+        self.As = matrix(inv_zf * (system.dae.fx - system.dae.fy * gyx))
         return self.As
 
     def calc_eigvals(self):
@@ -96,6 +113,16 @@ class EIG(BaseRoutine):
 
         return self.mu, self.part_fact
 
+    def summary(self):
+        """
+        Print out a summary to ``logger.info``.
+        """
+        out = list()
+        out.append('')
+        out.append('-> Eigenvalue Analysis:')
+        out_str = '\n'.join(out)
+        logger.info(out_str)
+
     def run(self, **kwargs):
         ret = False
         system = self.system
@@ -105,21 +132,24 @@ class EIG(BaseRoutine):
             return ret
         else:
             if system.TDS.initialized is False:
-                system.TDS._initialize()
-                system.TDS._implicit_step()
+                system.TDS.init()
 
         if system.dae.n == 0:
             logger.warning('No dynamic model. Eig analysis will not continue.')
             return ret
 
+        if sum(system.dae.zf != 0) != len(system.dae.zf):
+            logger.error("System contains zero time constant. Eigenvalue analysis cannot continue.")
+            return ret
+
+        self.summary()
         t1, s = elapsed()
-        logger.info('-> Eigenvalue Analysis:')
 
         self.calc_state_matrix()
         self.calc_part_factor()
 
         if not self.system.files.no_output:
-            self.write_report()
+            self.report()
             if system.options.get('state_matrix') is True:
                 self.export_state_matrix()
 
@@ -189,23 +219,26 @@ class EIG(BaseRoutine):
 
     def export_state_matrix(self):
         """
-        Export As to a ``<CaseName>_As.mat`` file.
+        Export state matrix to a ``<CaseName>_As.mat`` file with the variable name ``As``, where
+        ``<CaseName>`` is the test case name.
+
+        State variable names are stored in variables ``x_name`` and ``x_tex_name``.
 
         Returns
         -------
         bool
             True if successful
-
         """
         system = self.system
         out = {'As': self.As,
                'x_name': np.array(system.dae.x_name, dtype=np.object),
+               'x_tex_name': np.array(system.dae.x_tex_name, dtype=np.object),
                }
         scipy.io.savemat(system.files.mat, mdict=out)
-        logger.info(f'State matrix saved to <{system.files.mat}>.')
+        logger.info(f'State matrix saved to "{system.files.mat}".')
         return True
 
-    def write_report(self):
+    def report(self):
         """
         Save eigenvalue analysis reports
 
@@ -215,7 +248,7 @@ class EIG(BaseRoutine):
         """
         system = self.system
         mu = self.mu
-        partfact = self.part_fact
+        part_fact = self.part_fact
 
         text = []
         header = []
@@ -225,9 +258,9 @@ class EIG(BaseRoutine):
         neig = len(mu)
         mu_real = mu.real()
         mu_imag = mu.imag()
-        npositive = sum(1 for x in mu_real if x > 0)
-        nzero = sum(1 for x in mu_real if x == 0)
-        nnegative = sum(1 for x in mu_real if x < 0)
+        n_positive = sum(1 for x in mu_real if x > 0)
+        n_zeros = sum(1 for x in mu_real if x == 0)
+        n_negative = sum(1 for x in mu_real if x < 0)
 
         numeral = []
         for idx, item in enumerate(range(neig)):
@@ -239,7 +272,7 @@ class EIG(BaseRoutine):
                 marker = ''
             numeral.append('#' + str(idx + 1) + marker)
 
-        # compute frequency, undamped frequency and damping
+        # compute frequency, un-damped frequency and damping
         freq = [0] * neig
         ufreq = [0] * neig
         damping = [0] * neig
@@ -256,7 +289,7 @@ class EIG(BaseRoutine):
         # obtain most associated variables
         var_assoc = []
         for prow in range(neig):
-            temp_row = partfact[prow, :]
+            temp_row = part_fact[prow, :]
             name_idx = list(temp_row).index(max(temp_row))
             var_assoc.append(system.dae.x_name[name_idx])
 
@@ -264,7 +297,7 @@ class EIG(BaseRoutine):
         for prow in range(neig):
             temp_row = []
             for pcol in range(neig):
-                temp_row.append(round(partfact[prow, pcol], 5))
+                temp_row.append(round(part_fact[prow, pcol], 5))
             pf.append(temp_row)
 
         text.append('')
@@ -275,7 +308,7 @@ class EIG(BaseRoutine):
         text.append('STATISTICS\n')
         header.append([''])
         rowname.append(['Positives', 'Zeros', 'Negatives'])
-        data.append([npositive, nzero, nnegative])
+        data.append([n_positive, n_zeros, n_negative])
 
         text.append('EIGENVALUE DATA\n')
         header.append([
@@ -288,18 +321,18 @@ class EIG(BaseRoutine):
              list(mu_real),
              list(mu_imag), ufreq, freq, damping])
 
-        cpb = 7  # columns per block
-        nblock = int(ceil(neig / cpb))
+        n_cols = 7  # columns per block
+        n_block = int(ceil(neig / n_cols))
 
-        if nblock <= 100:
-            for idx in range(nblock):
-                start = cpb * idx
-                end = cpb * (idx + 1)
+        if n_block <= 100:
+            for idx in range(n_block):
+                start = n_cols * idx
+                end = n_cols * (idx + 1)
                 text.append('PARTICIPATION FACTORS [{}/{}]\n'.format(
-                    idx + 1, nblock))
+                    idx + 1, n_block))
                 header.append(numeral[start:end])
                 rowname.append(system.dae.x_name)
                 data.append(pf[start:end])
 
         dump_data(text, header, rowname, data, system.files.eig)
-        logger.info(f'Report saved to <{system.files.eig}>.')
+        logger.info(f'Report saved to "{system.files.eig}".')